添加俱乐部

+++ GongYunTong/Classes/Login/Controller/VHSLoginController.m
@@ -213,7 +213,7 @@
             VHSActionData *ac = [[VHSActionData alloc] init];
             ac.actionId = [VHSCommon getTimeStamp];
             ac.memberId = net.memberId;
-            ac.step = [NSString stringWithFormat:@"%ld", net.step];
+            ac.step = [NSString stringWithFormat:@"%ld", (long)net.step];
             ac.recordTime = net.sportDate;
             ac.macAddress = net.handMac;
             ac.actionType = [net.handMac isEqualToString:@"0"] ? @"2" : @"1";
